Movement Refinement and General Cleanup
Removed two out of date tank minigame scenes and a "scrip" folder with unused script in it. On the rigidBody movement script, the player now cannot turn and move at the same time. The character movement script added in previous commit is broken, dunno why, couldn't be bothered to fix it with rigidbody script working fine.