Shell projectile made, script in the works
Finished making a projectile object in unit that uses a particle effect for the fire around the shell and a trail object as it fires. The script for spawn the object is getting an error and the firing script doesn't work as intended. Supposedly both scripts were suppose to make the tank fire the projectile when clicking the mouse button.